+++ react-native.config.js
@@ -1,8 +1,7 @@
 module.exports = {
-    project: {
-      ios: {},
-      android: {},
-    },
-    assets: ['./src/fonts/'],
-  };
-  
+  project: {
+    ios: {},
+    android: {},
+  },
+  assets: ['./src/fonts/'],
+}
