+++ src/screens/CreationScreen.jsx
@@ -3,7 +3,7 @@ import { View, ScrollView, StatusBar } from 'react-native'
 import useApi from '../hooks/useApi/useApi'
 import SliderContainer from '../components/SliderContainer'
 import CommentsComponent from '../components/commentsComponent'
-import { ModalComponent } from '../components/modalComponent'
+import ModalComponent from '../components/modalComponent'
 
 const CreationScreen = () => {
   const [response,, handleRequest] = useApi()
